+++ tests/05_provider_test.py
@@ -42,10 +42,10 @@ def p():
 
 
 def test_provider_status(p):
-    p.provider_status("online_provider", "https://www.google.com")
+    p.provider_status("https://www.google.com")
 
     with pytest.raises(ConnectionError):
-        p.provider_status("offline_provider", "https://www.thiswebsiteisoffline.nl/")
+        p.provider_status("https://www.thiswebsiteisoffline.nl/")
 
 
 def test_providerbase__init__(p):