+++ tests/07_provider_ucsc_test.py
@@ -2,8 +2,8 @@ import genomepy
 import gzip
 import os
 import pytest
-import shutil
 
+from shutil import copyfile
 from tempfile import TemporaryDirectory
 
 
@@ -69,16 +69,16 @@ def test_genome_info_tuple(p):
 
 def test_get_genome_download_link(p):
     link = p.get_genome_download_link("sacCer3", mask="soft")
-    assert (
-        link
-        == "http://hgdownload.soe.ucsc.edu/goldenPath/sacCer3/bigZips/chromFa.tar.gz"
-    )
+    assert link in [
+        "http://hgdownload.soe.ucsc.edu/goldenPath/sacCer3/bigZips/chromFa.tar.gz",
+        "http://hgdownload.soe.ucsc.edu/goldenPath/sacCer3/bigZips/sacCer3.fa.gz",
+    ]
 
     link = p.get_genome_download_link("danRer7", mask="hard")
-    assert (
-        link
-        == "http://hgdownload.soe.ucsc.edu/goldenPath/danRer7/bigZips/danRer7.fa.masked.gz"
-    )
+    assert link in [
+        "http://hgdownload.soe.ucsc.edu/goldenPath/danRer7/bigZips/chromFaMasked.tar.gz",
+        "http://hgdownload.soe.ucsc.edu/goldenPath/danRer7/bigZips/danRer7.fa.masked.gz",
+    ]
 
 
 def test__post_process_download(p):
@@ -98,7 +98,7 @@ def test__post_process_download(p):
 
         # copy fa file for unmasking
         g = os.path.join(tmpdir, localname + ".fa")
-        shutil.copyfile("tests/data/gap.fa", g)
+        copyfile("tests/data/gap.fa", g)
 
         p._post_process_download(
             name=None, localname=localname, out_dir=tmpdir, mask="none"