+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - added retries to UCSC and NCBI
 - added retries to Travis tests
 - Bucketcache improvements
+- `genomepy search` will keep searching after an exact match is found
 
 ### Fixed