+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 - `genomepy clean` wont complain when there is nothing to clean
 - properly gzip the annotation.gtf if it was unzipped during sanitizing.
+- `genomepy install` can use the URL provider again
 
 ## [0.9.0] - 2020-09-01